requirements.txt
ipykernel
ipython
ipywidgets
ipympl
jupyter
jupyterlab
nodejs
matplotlib
notebook
numpy
pandas
pickleshare
pillow
torch
torchvision
seaborn
jupyter_contrib_nbextensions 
jupyter_nbextensions_configurator